Use Span<T>.Slice check

+++ src/libraries/System.Private.CoreLib/src/System/Text/StringBuilder.cs
@@ -662,7 +662,10 @@ namespace System.Text
             char[] chunkChars = m_ChunkChars;
             int chunkLength = m_ChunkLength;
 
-            if (((uint)chunkLength + (uint)repeatCount) <= (uint)chunkChars.Length)
+            // Try to fit the whole repeatCount in the current chunk
+            // Use the same check as Span<T>.Slice for 64-bit so it can be folded
+            // Since repeatCount can't be negative, there's no risk for it to overflow on 32 bit
+            if (((nuint)(uint)chunkLength + (nuint)(uint)repeatCount) <= (nuint)(uint)chunkChars.Length)
             {
                 chunkChars.AsSpan(chunkLength, repeatCount).Fill(value);
                 m_ChunkLength += repeatCount;
